sui_package_dump

Function dump

pub async fn dump(
    rpc_url: String,
    output_dir: PathBuf,
    before_checkpoint: Option<u64>,
) -> Result<()>

Ensure all packages created before before_checkpoint are written to the output_directory, from the GraphQL service at rpc_url.

output_dir can be a path to a non-existent directory, an existing empty directory, or an existing directory written to in the past. If the path is non-existent, the invocation creates it. If the path exists but is empty, the invocation writes to the directory. If the directory has been written to in the past, the invocation picks back up where the previous invocation left off.